+++ packages/apollo-link-error/src/__tests__/index.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 import { ApolloLink, execute, Observable } from 'apollo-link';
-
 import { onError, ErrorLink } from '../';
+import { ServerError, throwServerError } from 'apollo-link-http-common';
 
 describe('error handling', () => {
   it('has an easy way to handle GraphQL errors', done => {
@@ -86,6 +86,42 @@ describe('error handling', () => {
     const mockLink = new ApolloLink(operation => {
       return new Observable(obs => {
         throw new Error('app is crashing');
+      });
+    });
+
+    const link = errorLink.concat(mockLink);
+
+    execute(link, { query }).subscribe({
+      error: e => {
+        expect(e.message).toBe('app is crashing');
+        expect(called).toBe(true);
+        done();
+      },
+    });
+  });
+  it('captures networkError.statusCode within links', done => {
+    const query = gql`
+      query Foo {
+        foo {
+          bar
+        }
+      }
+    `;
+
+    let called;
+    const errorLink = onError(({ operation, networkError }) => {
+      expect(networkError.message).toBe('app is crashing');
+      expect(networkError.name).toBe('ServerError');
+      expect((networkError as ServerError).statusCode).toBe(500);
+      expect((networkError as ServerError).response.ok).toBe(false);
+      expect(operation.operationName).toBe('Foo');
+      called = true;
+    });
+
+    const mockLink = new ApolloLink(operation => {
+      return new Observable(obs => {
+        const response = { status: 500, ok: false } as Response;
+        throwServerError(response, 'ServerError', 'app is crashing');
       });
     });
 
@@ -195,6 +231,7 @@ describe('error handling', () => {
 
     setTimeout(done, 10);
   });
+
   it('includes the operation and any data along with a graphql error', done => {
     const query = gql`
       query Foo {
@@ -388,5 +425,190 @@ describe('error handling with class', () => {
     sub.unsubscribe();
 
     setTimeout(done, 10);
+  });
+});
+
+describe('support for request retrying', () => {
+  const QUERY = gql`
+    query Foo {
+      foo {
+        bar
+      }
+    }
+  `;
+  const ERROR_RESPONSE = {
+    errors: [
+      {
+        name: 'something bad happened',
+        message: 'resolver blew up',
+      },
+    ],
+  };
+  const GOOD_RESPONSE = {
+    data: { foo: true },
+  };
+  const NETWORK_ERROR = {
+    message: 'some other error',
+  };
+
+  it('returns the retried request when forward(operation) is called', done => {
+    let errorHandlerCalled = false;
+
+    let timesCalled = 0;
+    const mockHttpLink = new ApolloLink(operation => {
+      if (timesCalled === 0) {
+        timesCalled++;
+        // simulate the first request being an error
+        return new Observable(observer => {
+          observer.next(ERROR_RESPONSE);
+          observer.complete();
+        });
+      } else {
+        return new Observable(observer => {
+          observer.next(GOOD_RESPONSE);
+          observer.complete();
+        });
+      }
+    });
+
+    const errorLink = new ErrorLink(
+      ({ graphQLErrors, response, operation, forward }) => {
+        try {
+          if (graphQLErrors) {
+            errorHandlerCalled = true;
+            expect(graphQLErrors).toEqual(ERROR_RESPONSE.errors);
+            expect(response.data).not.toBeDefined();
+            expect(operation.operationName).toBe('Foo');
+            expect(operation.getContext().bar).toBe(true);
+            // retry operation if it resulted in an error
+            return forward(operation);
+          }
+        } catch (error) {
+          done.fail(error);
+        }
+      },
+    );
+
+    const link = errorLink.concat(mockHttpLink);
+
+    execute(link, { query: QUERY, context: { bar: true } }).subscribe({
+      next(result) {
+        try {
+          expect(errorHandlerCalled).toBe(true);
+          expect(result).toEqual(GOOD_RESPONSE);
+        } catch (error) {
+          return done.fail(error);
+        }
+      },
+      complete() {
+        done();
+      },
+    });
+  });
+
+  it('supports retrying when the initial request had networkError', done => {
+    let errorHandlerCalled = false;
+
+    let timesCalled = 0;
+    const mockHttpLink = new ApolloLink(operation => {
+      if (timesCalled === 0) {
+        timesCalled++;
+        // simulate the first request being an error
+        return new Observable(observer => {
+          observer.error(NETWORK_ERROR);
+        });
+      } else {
+        return new Observable(observer => {
+          observer.next(GOOD_RESPONSE);
+          observer.complete();
+        });
+      }
+    });
+
+    const errorLink = new ErrorLink(
+      ({ networkError, response, operation, forward }) => {
+        try {
+          if (networkError) {
+            errorHandlerCalled = true;
+            expect(networkError).toEqual(NETWORK_ERROR);
+            return forward(operation);
+          }
+        } catch (error) {
+          done.fail(error);
+        }
+      },
+    );
+
+    const link = errorLink.concat(mockHttpLink);
+
+    execute(link, { query: QUERY, context: { bar: true } }).subscribe({
+      next(result) {
+        try {
+          expect(errorHandlerCalled).toBe(true);
+          expect(result).toEqual(GOOD_RESPONSE);
+        } catch (error) {
+          return done.fail(error);
+        }
+      },
+      complete() {
+        done();
+      },
+    });
+  });
+
+  it('returns errors from retried requests', done => {
+    let errorHandlerCalled = false;
+
+    let timesCalled = 0;
+    const mockHttpLink = new ApolloLink(operation => {
+      if (timesCalled === 0) {
+        timesCalled++;
+        // simulate the first request being an error
+        return new Observable(observer => {
+          observer.next(ERROR_RESPONSE);
+          observer.complete();
+        });
+      } else {
+        return new Observable(observer => {
+          observer.error(NETWORK_ERROR);
+        });
+      }
+    });
+
+    const errorLink = new ErrorLink(
+      ({ graphQLErrors, networkError, response, operation, forward }) => {
+        try {
+          if (graphQLErrors) {
+            errorHandlerCalled = true;
+            expect(graphQLErrors).toEqual(ERROR_RESPONSE.errors);
+            expect(response.data).not.toBeDefined();
+            expect(operation.operationName).toBe('Foo');
+            expect(operation.getContext().bar).toBe(true);
+            // retry operation if it resulted in an error
+            return forward(operation);
+          }
+        } catch (error) {
+          done.fail(error);
+        }
+      },
+    );
+
+    const link = errorLink.concat(mockHttpLink);
+
+    let observerNextCalled = false;
+    execute(link, { query: QUERY, context: { bar: true } }).subscribe({
+      next(result) {
+        // should not be called
+        observerNextCalled = true;
+      },
+      error(error) {
+        // note that complete will not be after an error
+        // therefore we should end the test here with done()
+        expect(errorHandlerCalled).toBe(true);
+        expect(observerNextCalled).toBe(false);
+        expect(error).toEqual(NETWORK_ERROR);
+        done();
+      },
+    });
   });
 });