+++ packages/apollo-link-schema/src/schemaLink.ts
@@ -2,6 +2,10 @@ import { ApolloLink, Operation, FetchResult, Observable } from 'apollo-link';
 import { execute, GraphQLSchema } from 'graphql';
 
 export namespace SchemaLink {
+  export type ResolverContextFunction = (
+    operation: Operation,
+  ) => Record<string, any>;
+
   export interface Options {
     /**
      * The schema to generate responses from.
@@ -16,14 +20,14 @@ export namespace SchemaLink {
     /**
      * A context to provide to resolvers declared within the schema.
      */
-    context?: any;
+    context?: ResolverContextFunction | Record<string, any>;
   }
 }
 
 export class SchemaLink extends ApolloLink {
   public schema: GraphQLSchema;
   public rootValue: any;
-  public context: any;
+  public context: SchemaLink.ResolverContextFunction | any;
 
   constructor({ schema, rootValue, context }: SchemaLink.Options) {
     super();
@@ -40,7 +44,9 @@ export class SchemaLink extends ApolloLink {
           this.schema,
           operation.query,
           this.rootValue,
-          this.context,
+          typeof this.context === 'function'
+            ? this.context(operation)
+            : this.context,
           operation.variables,
           operation.operationName,
         ),