+++ packages/apollo-link-batch/CHANGELOG.md
@@ -1,6 +1,22 @@
 # Change log
 
-### vNext
+----
+
+**NOTE:** This changelog is no longer maintained. Changes are now tracked in
+the top level [`CHANGELOG.md`](https://github.com/apollographql/apollo-link/blob/master/CHANGELOG.md).
+
+----
+
+### 1.1.5
+
+- No changes
+
+### 1.1.4
+- Added `graphql` 14 to peer and dev deps; Updated `@types/graphql` to 14  <br/>
+  [@hwillson](http://github.com/hwillson) in [#789](https://github.com/apollographql/apollo-link/pull/789)
+
+### 1.1.3
+- Update apollo-link [#559](https://github.com/apollographql/apollo-link/pull/559)
 
 ### 1.1.2
 - Fix typing of Operation parameters [PR#525](https://github.com/apollographql/apollo-link/pull/525)