+++ packages/apollo-link-schema/CHANGELOG.md
@@ -1,9 +1,26 @@
 # Change log
 
-### vNEXT
+----
+
+**NOTE:** This changelog is no longer maintained. Changes are now tracked in
+the top level [`CHANGELOG.md`](https://github.com/apollographql/apollo-link/blob/master/CHANGELOG.md).
+
+----
+
+### 1.1.2
+
+- No changes
+
+### 1.1.1
+- Added `graphql` 14 to peer and dev deps; Updated `@types/graphql` to 14  <br/>
+  [@hwillson](http://github.com/hwillson) in [#789](https://github.com/apollographql/apollo-link/pull/789)
+
+### 1.1.0
+- Update apollo-link [#559](https://github.com/apollographql/apollo-link/pull/559)
+- allow context to be a function that accept an operation and returns a context for the resolvers [PR#363](https://github.com/apollographql/apollo-link/pull/363)
 
 ### 1.0.6
-- udate apollo link with zen-observable-ts [PR#515](https://github.com/apollographql/apollo-link/pull/515)
+- update apollo link with zen-observable-ts [PR#515](https://github.com/apollographql/apollo-link/pull/515)
 
 ### 1.0.5
 - Include sourcemap in bundle
@@ -23,4 +40,3 @@
 ### 1.0.0
 - support mocking
 - support server rendering
-